Moved lite members itests to it/litemembers package
